+++ godot/utils/automation/State.cs
@@ -6,7 +6,7 @@ namespace Lavos.Utils.Automation
     {
         public string Tag { get; set; }
         public Action Enter { get; set; }
-        public Action<float> Process { get; set; }
+        public Action<double> Process { get; set; }
         public Action Exit { get; set; }
     }
 }
